+++ lib/typesafe_enum/value_enum.rb
@@ -1,0 +1,9 @@
+module TypesafeEnum
+  class ValueEnum < Enum
+    def initialize(value)
+      @value = value
+    end
+
+    attr_reader :value
+  end
+end